+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Update instance header after overlaying source bibliographic record process. Fixes UIIN-3282.
 * Missing values in the version history modal for Suppressed from discovery and Staff suppressed fields. Fixed UIIN-3277.
 * Enable MARC-related options for shared MARC instance on member tenant. Fixes UIIN-3292.
+* ECS FOLIO instances: Inventory version history feature toggle/icon. Refs UIIN-3284.
 
 ## [13.0.0](https://github.com/folio-org/ui-inventory/tree/v13.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.12...v13.0.0)